Add requests, pageable and test for cntrllrs, srvcs and vldtr

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -1,8 +1,9 @@
 package ru.practicum.shareit.item.service;
 
 import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.booking.mapper.BookingMapper;
 import ru.practicum.shareit.booking.repository.BookingRepository;
@@ -15,11 +16,13 @@ import ru.practicum.shareit.item.model.Comment;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.item.repository.CommentRepository;
 import ru.practicum.shareit.item.repository.ItemRepository;
+import ru.practicum.shareit.request.repository.ItemRequestRepository;
 import ru.practicum.shareit.user.repository.UserRepository;
 import ru.practicum.shareit.validator.CustomValidator;
 
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -38,9 +41,10 @@ public class ItemServiceImpl implements ItemService {
 
     private final CommentRepository commentRepository;
 
+    private final ItemRequestRepository itemRequestRepository;
+
     private final CustomValidator customValidator;
 
-    @SneakyThrows
     @Transactional
     public ItemDto createItem(ItemDto itemDto, Long userId) {
         log.info("Попытка добавить новый предмет");
@@ -48,11 +52,13 @@ public class ItemServiceImpl implements ItemService {
         isUserExistAndAuthorizated(userId);
         Item item = ItemMapper.toItem(itemDto);
         item.setOwner(userRepository.getReferenceById(userId));
+        if (itemDto.getRequestId() != null) {
+            item.setRequest(itemRequestRepository.getReferenceById(itemDto.getRequestId()));
+        }
         log.info("Пользователь id: {} добавил новый предмет id: {}", userId, itemDto);
         return ItemMapper.toItemDto(itemRepository.save(item));
     }
 
-    @SneakyThrows
     @Transactional
     public ItemDto updateItem(ItemDto itemDto, Long itemId, Long userId) {
         log.info("Попытка обновить информацию о предмете id: {}", itemId);
@@ -62,7 +68,7 @@ public class ItemServiceImpl implements ItemService {
             log.info("Пользователь пытается обновить информацию о чужом предмете");
             throw new OwnerIdAndUserIdException("Обновлять информацию о предмете могут только владельцы");
         }
-        Item item = itemRepository.findById(itemId).get();
+        Item item = itemRepository.getReferenceById(itemId);
         if (itemDto.getName() != null) {
             item.setName(itemDto.getName());
         }
@@ -82,23 +88,23 @@ public class ItemServiceImpl implements ItemService {
         log.info("Пользователь id: {} пытается удалить предмет id: {}", itemId);
         isUserExistAndAuthorizated(userId);
         isItemExist(itemId);
+        if (!itemRepository.getOwnerId(itemId).equals(userId)) {
+            log.info("Пользователь пытается удалить чужой предмет");
+            throw new OwnerIdAndUserIdException("Удалять предметы могут только владельцы");
+        }
         itemRepository.deleteById(itemId);
         log.info("Предмет id: {} удален", itemId);
         return "Предмет удален";
     }
 
-    @SneakyThrows
     @Transactional
     public CommentDto addComment(CommentDto commentDto, Long itemId, Long userId) {
         log.info("Пользователь id: {} добавляет комментарий для предмета id: {}", userId, itemId);
         isUserExistAndAuthorizated(userId);
         isItemExist(itemId);
-        LocalDateTime createTime = LocalDateTime.now();
-        if (commentDto.getText().equals(null) || commentDto.getText().isEmpty()) {
-            log.info("Комментарий не содержит текст");
-            throw new EmptyCommentTextException("Текст комментария не может быть пустым");
-        }
-        if (bookingRepository.bookingsBeforeNowCount(itemId, userId, createTime) == 0) {
+        customValidator.isCommentValid(commentDto);
+        commentDto.setCreated(LocalDateTime.now());
+        if (bookingRepository.bookingsBeforeNowCount(itemId, userId, commentDto.getCreated()).equals(0L)) {
             log.info("Пользователь пока еще не брал в аренду предмет");
             throw new NoBookedYetException("Предмент не арендовался пользователем");
         }
@@ -106,13 +112,13 @@ public class ItemServiceImpl implements ItemService {
                 .author(userRepository.getReferenceById(userId))
                 .item(itemRepository.getReferenceById(itemId))
                 .text(commentDto.getText())
-                .created(createTime)
+                .created(commentDto.getCreated())
                 .build();
         log.info("Пользователь id: {} добавил комментарий для предмета id: {}", userId, itemId);
         return CommentMapper.toCommentDto(commentRepository.save(comment));
     }
 
-    @SneakyThrows
+    @Transactional
     public CommentDto updateComment(CommentDto commentDto, Long itemId, Long commentId, Long userId) {
         log.info("Пользователь id: {} пытается изменить комментарий id: {} для " +
                 "предмета id: {}", userId, commentId, itemId);
@@ -132,19 +138,22 @@ public class ItemServiceImpl implements ItemService {
         return CommentMapper.toCommentDto(commentRepository.save(comment));
     }
 
-    @SneakyThrows
+    @Transactional
     public String deleteComment(Long commentId, Long itemId, Long userId) {
         log.info("Пользователь id: {} пытается удалить комментарий id: {} для " +
                 "предмета id: {}", userId, commentId, itemId);
         isUserExistAndAuthorizated(userId);
         isItemExist(itemId);
         isCommentExist(commentId);
+        if (!commentRepository.getCommentAuthorId(commentId).equals(userId)) {
+            log.info("Пользователь id: {} не является автором комментария id: {}", userId, commentId);
+            throw new OwnerIdAndUserIdException("Только авторы комментария могут его редактировать");
+        }
         log.info("Комментарий удален");
         commentRepository.deleteById(commentId);
         return "Комментарий удален";
     }
 
-    @SneakyThrows
     @Transactional
     public ItemDto findItemById(Long userId, Long itemId) {
         log.info("Попытка получить информацию о предмете id:{}", itemId);
@@ -159,30 +168,32 @@ public class ItemServiceImpl implements ItemService {
         itemDto.setComments(getItemComments(itemId));
         if (item.getOwner().getId().equals(userId)) {
             log.info("Получена информация о прошлом и следующем бронированиях предмета id: {}", itemId);
-            return getItemWithBookings(itemDto);
+            return getBookingsForItem(itemDto);
         }
         log.info("Получена информация о предмете id: {}", itemId);
         return itemDto;
     }
 
-    @SneakyThrows
     @Transactional
-    public List<ItemDto> findUserAllItems(Long userId) {
+    public List<ItemDto> findUserAllItems(Long userId, Integer from, Integer size) {
         log.info("Попытка получить список предметов пользователя {}", userId);
+        isUserExistAndAuthorizated(userId);
+        customValidator.isPageableParamsCorrect(from, size);
+        Pageable pageRequest = PageRequest.of(from, size);
         log.info("Получен список предметов, размещенных пользователем {}", userId);
-        return itemRepository.findByOwner(userRepository.getReferenceById(userId)).stream()
+        return itemRepository.findByOwner(userRepository.getReferenceById(userId), pageRequest).stream()
                 .map(ItemMapper::toItemDto)
-                .map(this::getItemWithBookings)
+                .map(this::getBookingsForItem)
                 .collect(Collectors.toList());
     }
 
     private List<CommentDto> getItemComments(Long itemId) {
-        return commentRepository.getCommentsByItem_IdOrderByIdAsc(itemId).stream()
+        return commentRepository.getCommentsByItemIdOrderByIdAsc(itemId).stream()
                 .map(CommentMapper::toCommentDto)
                 .collect(Collectors.toList());
     }
 
-    private ItemDto getItemWithBookings(ItemDto itemDto) {
+    private ItemDto getBookingsForItem(ItemDto itemDto) {
         LocalDateTime nowTime = LocalDateTime.now();
         itemDto.setLastBooking(BookingMapper
                 .toBookingForItemDto(Optional.ofNullable(bookingRepository.getLastBooking(nowTime, itemDto.getId()))
@@ -193,22 +204,22 @@ public class ItemServiceImpl implements ItemService {
         return itemDto;
     }
 
-    @SneakyThrows
     @Transactional
-    public List<ItemDto> searchItemByNameAndDescription(String text) {
+    public List<ItemDto> searchItemByNameAndDescription(String text, Integer from, Integer size) {
         log.info("Попытка получить список предметов, доступных для аренды, по запросу {}", text);
+        customValidator.isPageableParamsCorrect(from, size);
         if (text == null || text.isBlank()) {
             log.info("В поисковой строке пусто, получен пустой список предметов");
             return Collections.emptyList();
         }
+        Pageable pageRequest = PageRequest.of(from, size);
         log.info("Получен список предметов, доступных для аренды, по ключевому слову {}", text);
-        return itemRepository.searchByText(text).stream()
+        return itemRepository.searchByText(text, pageRequest).stream()
                 .map(ItemMapper::toItemDto)
                 .filter(o -> o.getAvailable() == true)
                 .collect(Collectors.toList());
     }
 
-    @SneakyThrows
     private void isUserExistAndAuthorizated(Long userId) {
         if (userId <= 0) {
             log.info("Пользователь не авторизован");
@@ -220,7 +231,6 @@ public class ItemServiceImpl implements ItemService {
         }
     }
 
-    @SneakyThrows
     private void isItemExist(Long itemId) {
         if (!itemRepository.existsById(itemId)) {
             log.info("Предмет с id:{} не найден", itemId);
@@ -228,7 +238,6 @@ public class ItemServiceImpl implements ItemService {
         }
     }
 
-    @SneakyThrows
     private void isCommentExist(Long commentId) {
         if (!commentRepository.existsById(commentId)) {
             log.info("Комментария id: {} не существует", commentId);
